requirements.txt
accelerate
bitsandbytes
decord
litserve
openai
peft
Pillow
sentencepiece
timm
torch
transformers